refactor: mv amp/__init__.py -> amp/base.py
